shipping/pyproject.toml
[project]
authors = []
dependencies = [
  "strawberry-graphql",
  "fastapi",
  "uvicorn",
]
description = ""
name = "api"
requires-python = ">=3.7"
version = "0.1.0"
[project.optional-dependencies]

[tool.black]
line-length = 88

[tool.isort]
combine_as_imports = true
combine_star = true
indent = 4
known_pytest = ["pytest"]
known_strawberry = ["strawberry"]
lines_after_imports = 2
profile = "black"
sections = ["FUTURE", "STDLIB", "PYTEST", "THIRDPARTY", "STRAWBERRY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
plugins = ["strawberry.ext.mypy_plugin"]

[tool.pdm]
[tool.pdm.scripts]
prod-server = "uvicorn main:app --host 0.0.0.0 --port 8080"
server = "uvicorn main:app --reload"

[tool.pdm.dev-dependencies]
dev = [
  "mypy>=0.982",
  "black>=22.10.0",
  "flake8>=5.0.4",
]

[build-system]
build-backend = "pdm.pep517.api"
requires = ["pdm-pep517>=1.0.0"]
